+++ src/App.tsx
@@ -2,10 +2,18 @@ import React from 'react';
 import { useRoutes } from 'hookrouter';
 import { routes } from './routes';
 import NotFoundPage from './pages/NotFound';
+import Header from './components/Header';
 
 const App = () => {
   const match = useRoutes(routes);
-  return match || <NotFoundPage />;
+  return match ? (
+    <>
+      <Header />
+      {match}
+    </>
+  ) : (
+    <NotFoundPage />
+  );
 };
 
 export default App;
