+++ src/components/Header/index.tsx
@@ -34,4 +34,4 @@ const Header = () => {
   );
 };
 
-export default Header;
+export default React.memo(Header);
