+++ src/components/PokemonCard/index.tsx
@@ -1,33 +1,9 @@
 import React, { FC } from 'react';
 import Heading from '../Heading';
 
+import { IPokemonCard } from './interface';
+
 import s from './PokemonCard.module.scss';
-
-export interface ICardItem {
-  name_clean: string;
-  abilities: string[];
-  stats: {
-    hp: number;
-    attack: number;
-    defense: number;
-    'special-attack': number;
-    'special-defense': number;
-    speed: number;
-  };
-  types: string[];
-  img: string;
-  name: string;
-  base_experience: number;
-  height: number;
-  id: number;
-  is_default: boolean;
-  order: number;
-  weight: number;
-}
-
-export interface IPokemonCard {
-  item: ICardItem;
-}
 
 const PokemonCard: FC<IPokemonCard> = ({ item }) => {
   return (
