+++ src/components/PokemonCard/interface.ts
@@ -1,0 +1,25 @@
+export interface ICardItem {
+  name_clean: string;
+  abilities: string[];
+  stats: {
+    hp: number;
+    attack: number;
+    defense: number;
+    'special-attack': number;
+    'special-defense': number;
+    speed: number;
+  };
+  types: string[];
+  img: string;
+  name: string;
+  base_experience: number;
+  height: number;
+  id: number;
+  is_default: boolean;
+  order: number;
+  weight: number;
+}
+
+export interface IPokemonCard {
+  item: ICardItem;
+}
