+++ src/pages/Empty/index.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import Header from '../../components/Header';
 
 import s from './Empty.module.scss';
 
@@ -9,7 +8,6 @@ interface IEmptyPage {
 const EmptyPage: React.FC<IEmptyPage> = ({ title }) => {
   return (
     <div className={s.root}>
-      <Header />
       <div>Empty page for {title}</div>
     </div>
   );
