+++ src/pages/Home/index.tsx
@@ -1,8 +1,6 @@
 import React from 'react';
 import { navigate } from 'hookrouter';
 
-import Header from '../../components/Header';
-
 import s from './Home.module.scss';
 import Button from '../../components/Button';
 import Layout from '../../components/Layout';
@@ -14,7 +12,6 @@ import Heading from '../../components/Heading';
 const HomePage = () => {
   return (
     <div className={s.root}>
-      <Header />
       <Layout className={s.contentWrap}>
         <div className={s.leftContent}>
           <div className={s.contentText}>
@@ -34,8 +31,9 @@ const HomePage = () => {
             onClick={() => navigate('/pokedex')}
             width={BUTTON_WIDTHS.NORMAL}
             size={BUTTON_SIZE.SMALL}
+            className=""
             color={BUTTON_COLORS.GREEN}>
-            See pokemons
+            <>See pokemons</>
           </Button>
         </div>
         <Parallax />
