+++ src/pages/Pokedex/index.tsx
@@ -1,14 +1,14 @@
 import React from 'react';
 import pokemons from './constants';
-import PokemonCard, { ICardItem } from '../../components/PokemonCard';
-import Header from '../../components/Header';
+import PokemonCard from '../../components/PokemonCard';
 
 import s from './Pokedex.module.scss';
+
+import { ICardItem } from '../../components/PokemonCard/interface';
 
 function PokedexPage() {
   return (
     <div className={s.root}>
-      <Header />
       <div className={s.content}>
         {(pokemons as ICardItem[]).map((item: ICardItem) => (
           <PokemonCard key={item.name_clean} item={item} />
