+++ src/utils.ts
@@ -1,0 +1,1 @@
+export type ValueOf<T> = T[keyof T];
